Updated convert membership
For the first table (recommended upgrade), members should appear if their total amount spent (not including tax) is equal to or greater than 55 ( the price difference between a regular and an executive member). The threshold is $2750 (without taxes) because 2% of this is $55. If they spend 2750 (not including taxes) they should qualify  they should upgrade to executive. If they spend anything less like 2749.99 (without taxes) they should not be eligible. I tested the boundaries and it works. Try to test it purchasing something on a regular member for 2749.99 and one for 2750 and it should work.

Before it would only populate if the rebate amount was more than $55, but since regular members wouldn't ever receive a rebate amount the condition would have never been met.